[move] パッケージ構造を動かす
