+++ src/main/scala/com/github/unchama/itemmigration/ItemMigrationConfiguration.scala
@@ -1,6 +1,8 @@
 package com.github.unchama.itemmigration
 
 import cats.effect.Bracket
+import com.github.unchama.itemmigration.domain.{ItemMigration, ItemMigrationSeq, ItemMigrationTarget}
+import com.github.unchama.itemmigration.service.ItemMigrationPersistence
 
 case class ItemMigrationConfiguration[F[_], T <: ItemMigrationTarget[F]](migrationSeq: ItemMigrationSeq,
                                                                          migrationTarget: T,
