+++ src/main/scala/com/github/unchama/itemmigration/domain/ItemMigrationTarget.scala
@@ -1,6 +1,6 @@
-package com.github.unchama.itemmigration
+package com.github.unchama.itemmigration.domain
 
-import com.github.unchama.itemmigration.ItemMigration.ItemStackConversion
+import com.github.unchama.itemmigration.domain.ItemMigration.ItemStackConversion
 
 /**
  * アイテムマイグレーションを行う対象。
