+++ src/main/scala/com/github/unchama/itemmigration/service/ItemMigrationPersistence.scala
@@ -1,8 +1,9 @@
-package com.github.unchama.itemmigration
+package com.github.unchama.itemmigration.service
 
 import cats.Monad
 import cats.effect.Resource
-import com.github.unchama.itemmigration.ItemMigration.VersionNumber
+import com.github.unchama.itemmigration.domain.ItemMigration
+import com.github.unchama.itemmigration.domain.ItemMigration.VersionNumber
 
 trait ItemMigrationPersistence[F[_], -T] {
   implicit val fMonad: Monad[F]
