+++ src/main/scala/com/github/unchama/seichiassist/infrastructure/migration/SeichiAssistPersistedItems.scala
@@ -1,8 +1,8 @@
 package com.github.unchama.seichiassist.infrastructure.migration
 
 import cats.effect.IO
-import com.github.unchama.itemmigration.ItemMigration.ItemStackConversion
-import com.github.unchama.itemmigration.ItemMigrationTarget
+import com.github.unchama.itemmigration.domain.ItemMigration.ItemStackConversion
+import com.github.unchama.itemmigration.domain.ItemMigrationTarget
 
 object SeichiAssistPersistedItems extends ItemMigrationTarget[IO] {
