+++ src/main/scala/com/github/unchama/seichiassist/infrastructure/migration/persistenceproviders/PersistedItemMigrationPersistence.scala
@@ -1,7 +1,7 @@
 package com.github.unchama.seichiassist.infrastructure.migration.persistenceproviders
 
 import cats.effect.IO
-import com.github.unchama.itemmigration.ItemMigrationPersistence
+import com.github.unchama.itemmigration.service.ItemMigrationPersistence
 import com.github.unchama.seichiassist.infrastructure.migration.SeichiAssistPersistedItems
 
 /**
