+++ src/main/scala/com/github/unchama/seichiassist/itemmigration/SeichiAssistWorldLevelData.scala
@@ -1,8 +1,8 @@
 package com.github.unchama.seichiassist.itemmigration
 
 import cats.effect.IO
-import com.github.unchama.itemmigration.ItemMigration.ItemStackConversion
-import com.github.unchama.itemmigration.ItemMigrationTarget
+import com.github.unchama.itemmigration.domain.ItemMigration.ItemStackConversion
+import com.github.unchama.itemmigration.domain.ItemMigrationTarget
 import com.github.unchama.itemmigration.target.WorldLevelData
 import com.github.unchama.seichiassist.SeichiAssist
 import com.github.unchama.util.external.{ExternalPlugins, ExternalServices}
